+++ .scalafmt.conf
@@ -10,3 +10,4 @@ docstrings = JavaDoc
 newlines.afterCurlyLambda = preserve
 docstrings.style = Asterisk
 docstrings.oneline = unfold
+project.excludeFilters = [ "core/src/main/scala-3.x" ]
