+++ tests/src/test/scala/cats/tests/OptionTSuite.scala
@@ -536,7 +536,7 @@ class OptionTSuite extends CatsSuite {
     Traverse[OptionT[List, *]]
 
     implicit val T: Traverse[ListWrapper] = ListWrapper.traverse
-    implicit val M: Monad[ListWrapper] = ListWrapper.monad
+    //implicit val M: Monad[ListWrapper] = ListWrapper.monad
     Functor[OptionT[ListWrapper, *]]
   }
